requirements-test.txt
pytest
pytest-flask==1.2.0
pytest-html
coverage
pytest-cov